requirements.txt
# pip install protobuf==3.20.*
langchain==0.0.116
openai==0.27.2
google-cloud-bigquery==3.7.0
jinja2==3.1.2
pandas
python-dotenv
db-dtypes
protobuf==3.20.*